max body size

+++ src/Webserv/VirtualServer.hpp
@@ -23,7 +23,7 @@ class VirtualServer
 		std::time_t		_keepalive_time;
 		std::time_t		_keepalive_timeout;
 		int				_cgi_timeout;
-		int				_max_body_size;
+		size_t			_max_body_size;
 
 		std::map<int, std::string>		_error_pages;
 		std::vector<std::string>		_indexes;
@@ -54,7 +54,7 @@ class VirtualServer
 		std::time_t		getKeepaliveTime(void)						const	{return _keepalive_time;}
 		std::time_t		getKeepaliveTimeout(void)					const	{return _keepalive_timeout;}
 		int				getCGITimeout()								const	{return _cgi_timeout;}	
-		long			getMaxBodySize()							const	{return _max_body_size;}
+		size_t			getMaxBodySize()							const	{return _max_body_size;}
 		std::map<int, std::string>		const&	getErrorPages()		const	{return _error_pages;}
 		std::vector<std::string>		const&	getIndexes(void)	const	{return _indexes;}
 		std::map<std::string, Location>	const&	getLocations(void)				const	{return _locations;}
